fts: interpret semicolon-separated attr values as value lists, performace
Often we use semicolons to separate list values in attribute value
strings. Postgres treats them as one word (and fails sometimes because
they are too long), but we want to search for single list values in full text search.
This is solved by simply replacing all ; by spaces in the full text
search specification for attributes.
This may lead to problems in some cases but we must live with that for the moment.
Someday, these "semicolon lists" should be converted to JSON arrays.

Improve search query performance by using a different method for finding
the nodes to search. Postgres chose a very slow plan when search results were
sorted by node id (the default).