utils/process.py wraps subprocess, overrides paths (fixes #774)
The new module utils.process provides functions
 * Popen
 * call
 * check_call
 * check_output
that resemble their counterparts in the subprocess module.
However, if an executable $NAME is called and there exists a
key $NAME in the "external" section of the mediatum config file,
then the name/path of the executable is replaced with the
value of said key before passing everything on to subprocess.

This allows to override executable file paths, e.g.
> [external]
> convert = /nix/store/blabla-imagemagick-1.2.3/bin/convert
ensures a specific version of convert is used.

This commit replaces all subprocess function invocations
with their respective utils.process counterparts,
with these notable exception:
 * utils/run_script_as_process.py
   contains a very general process calling interface,
   where silently overriding paths might lead to surprises.
   So here, it's the responsibility of the caller
   to resolve the name of the executable if needed.
 * lib/pdf/parsepdf.py
   has a function that calls this
   very module as own python process.
   There is no need to replace anything here.
 * bin/mysql_migrate.py
   is usually not called during
   normal mediatum operation.

Test cases are included,
together with a tiny shell script
to test actual redirection.