fulltext search: use outerjoin for fts table
Fixes problems with empty fts tables.
For example, AttributeMatch must work without entries in the fts table,
but didn't if an FullMatch / FulltextMatch occurred in the same query.
The old inner join threw away nodes that don't have an entry in the fts
table, which is not what we want. An outer join keeps these nodes.