edit, files module: show either change or add option for digital objects
Adding objects to existing objects doesn't make sense and creates an illegal state for the node.
Nodes may have only one "digital object" but we don't really enforce this.
This change removes the possibility for editors to create such a state in this module.
Changing objects only makes sense when something is present, which is quite obvious ;)

closes #877