add in-memory node caching for common singleton nodes
Home, Root and Collections are fetched in many places
but don't change during runtime (we assume).
Caching them in memory is a good idea but we must be careful about
detached objects. To make it safe, the object is merged with the current
session all time when requested. This is quite slow and should be done
only once per request.