HTTP cache duration can be configured in node classes and ContentBase classes now
This should only be used for things that are known to be public (independent of user, date and IP address).
The old way of setting cache headers leaked too much private content and fixing it would be costly (too much checking needed for little gain).

With current settings, errors for inaccessible nodes can be cached for 30s, the collections root for 600s.